requirements.txt
discord.py
aiohttp
wavelink==1.3.5
sponsorblock
loguru
python-dotenv
spotipy
tinydb